requirements.txt
fastapi
uvicorn
python-dotenv
requests
pillow
jinja2
python-multipart